# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algerian Agriculture Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  name: string;
  nameAr: string;
  icon: string;
  description: string;
  productCount: number;
}

export interface Wilaya {
  id: string;
  code: number;
  name: string;
  nameAr: string;
  region: string;
}

export interface Seller {
  id: string;
  name: string;
  wilayaId: string;
  wilayaName: string;
  phone: string;
  bio: string;
  rating: number;
  reviewCount: number;
  productCount: number;
  joinedAt: string;
  verified: boolean;
  profileImage?: string;
}

export interface Product {
  id: string;
  name: string;
  nameAr: string;
  description: string;
  categoryId: string;
  categoryName: string;
  sellerId: string;
  sellerName: string;
  wilayaId: string;
  wilayaName: string;
  price: number;
  unit: string;
  minOrder: number;
  stock: number;
  images: string[];
  featured: boolean;
  organic: boolean;
  rating: number;
  reviewCount: number;
  tags: string[];
  createdAt: string;
}

export interface CreateProduct {
  name: string;
  nameAr: string;
  description: string;
  categoryId: string;
  sellerId: string;
  wilayaId: string;
  price: number;
  unit: string;
  minOrder: number;
  stock: number;
  organic: boolean;
  tags?: string[];
}

export type MarketPriceTrend =
  (typeof MarketPriceTrend)[keyof typeof MarketPriceTrend];

export const MarketPriceTrend = {
  up: "up",
  down: "down",
  stable: "stable",
} as const;

export interface MarketPrice {
  id: string;
  productName: string;
  productNameAr: string;
  categoryName: string;
  priceMin: number;
  priceMax: number;
  unit: string;
  trend: MarketPriceTrend;
  changePercent: number;
  updatedAt: string;
  wilayaName: string;
}

export interface CreateMessage {
  senderName: string;
  senderEmail: string;
  senderPhone?: string;
  subject: string;
  body: string;
  recipientSellerId?: string;
}

export interface Message {
  id: string;
  senderName: string;
  senderEmail: string;
  senderPhone?: string;
  subject: string;
  body: string;
  recipientSellerId?: string;
  createdAt: string;
}

export type ListProductsParams = {
  categoryId?: string;
  wilayaId?: string;
  minPrice?: number;
  maxPrice?: number;
  search?: string;
  featured?: boolean;
  limit?: number;
  offset?: number;
};

export type ListProducts200 = {
  products: Product[];
  total: number;
};

export type ListCategories200 = {
  categories: Category[];
};

export type ListSellers200 = {
  sellers: Seller[];
};

export type ListWilayas200 = {
  wilayas: Wilaya[];
};

export type ListMarketPrices200 = {
  prices: MarketPrice[];
};