# lib/api-zod/src/generated/types/createMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algerian Agriculture Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMessage {
  senderName: string;
  senderEmail: string;
  senderPhone?: string;
  subject: string;
  body: string;
  recipientSellerId?: string;
}