# lib/api-zod/src/generated/types/createProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algerian Agriculture Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProduct {
  name: string;
  nameAr: string;
  description: string;
  categoryId: string;
  sellerId: string;
  wilayaId: string;
  price: number;
  unit: string;
  minOrder: number;
  stock: number;
  organic: boolean;
  tags?: string[];
}