# lib/api-zod/src/generated/types/marketPrice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algerian Agriculture Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketPriceTrend } from "./marketPriceTrend";

export interface MarketPrice {
  id: string;
  productName: string;
  productNameAr: string;
  categoryName: string;
  priceMin: number;
  priceMax: number;
  unit: string;
  trend: MarketPriceTrend;
  changePercent: number;
  updatedAt: string;
  wilayaName: string;
}