# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algerian Agriculture Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  nameAr: string;
  description: string;
  categoryId: string;
  categoryName: string;
  sellerId: string;
  sellerName: string;
  wilayaId: string;
  wilayaName: string;
  price: number;
  unit: string;
  minOrder: number;
  stock: number;
  images: string[];
  featured: boolean;
  organic: boolean;
  rating: number;
  reviewCount: number;
  tags: string[];
  createdAt: string;
}