# lib/api-zod/src/generated/types/seller.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Algerian Agriculture Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Seller {
  id: string;
  name: string;
  wilayaId: string;
  wilayaName: string;
  phone: string;
  bio: string;
  rating: number;
  reviewCount: number;
  productCount: number;
  joinedAt: string;
  verified: boolean;
  profileImage?: string;
}